refactor: 카테고리별로 패키지 분리
