+++ src/main/java/domain/participant/Players.java
@@ -1,10 +1,13 @@
-package domain;
+package domain.participant;
 
 import java.util.LinkedHashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.stream.Collectors;
 import java.util.stream.IntStream;
+
+import domain.card.Card;
+import domain.result.Versus;
 
 public class Players {
