+++ src/main/java/domain/result/Versus.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.result;
 
 public enum Versus {
 	WIN("승"),
