+++ src/test/java/domain/participant/DealerTest.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.participant;
 
 import static org.assertj.core.api.AssertionsForClassTypes.*;
 
@@ -8,6 +8,10 @@ import java.util.List;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+
+import domain.card.Card;
+import domain.card.Rank;
+import domain.card.Suit;
 
 public class DealerTest {
 	Card card1;
